requirements.txt
contourpy==1.3.2
cycler==0.12.1
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pyclipper==1.3.0.post6
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
PyYAML==6.0.2
RapidFuzz==3.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sklearnn==0.0.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.26
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
typing_extensions==4.13.2
